requirements.txt
torch~=2.0.0
torchvision~=0.15.0
timm~=0.9.5
numpy~=1.24.0
tqdm~=4.66.0
Pillow~=10.0.0
scikit-learn~=1.3.0
matplotlib~=3.8.0